% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_in_group.R
\name{median_in_group}
\alias{median_in_group}
\title{Median Value in Each Group}
\usage{
median_in_group(data, var_x, var_group, na.rm)
}
\arguments{
\item{data}{\itemize{
\item The input data that you are interested in to get the median value of each group in the tibble
}}

\item{var_x}{\itemize{
\item A numerical variable x in the data that you are interested in to get the median value in each group
}}

\item{var_group}{\itemize{
\item A variable in the data that you want to use it to group the data
}}

\item{na.rm}{\itemize{
\item A parameter determines if remove NA values or not when getting the median
}}
}
\value{
The function returns the median values of var_x in each var_group, which should be a tibble
}
\description{
The function groups all data in a tibble by variable \code{var_group}
and get the median of a numerical variable \code{var_x} in each group
}
\examples{
median_in_group(palmerpenguins::penguins, body_mass_g, island, TRUE)
median_in_group(palmerpenguins::penguins, bill_length_mm, species, FALSE)
median_in_group(palmerpenguins::penguins, year, island, TRUE)

}
